reafactor: useFilter로 모으기
