+++ app/component/main/MainContent/MainGridContainer.tsx
@@ -2,62 +2,25 @@
 
 import { getContents } from '@/app/api/content';
 import { ContentResponseType } from '@/app/type/content';
+import useFilter from '@/hooks/useFilter';
 import { useSearchParams } from 'next/navigation';
 import { useState, useEffect } from 'react';
 import Error from '../../common/Error';
 import Loading from '../../common/Loading';
 import ContentFilter from './ContentFilter';
 import { MainContentCard } from './MainContentCard';
-import { CategoryKey, Filter } from './category';
+import { CategoryKey } from './category';
 
 export type Topic = '성대한만남' | '기술교류회' | '공개형교육' | '성대한활동';
-
-const groupedFilterList: {
-  [key in Topic]: Filter[];
-} = {
-  성대한만남: [
-    { name: '성대의 성대한 특강', checked: false },
-    { name: '성대의 성대한 스토리', checked: false },
-    { name: '글로벌 IT전문가와 킹고인의 만남', checked: false },
-  ],
-  기술교류회: [{ name: 'S-TOP', checked: false }],
-  공개형교육: [
-    { name: '공개형 온라인 강의', checked: false },
-    { name: '공개형 교재', checked: false },
-  ],
-  성대한활동: [
-    { name: '글로벌 챌린지', checked: false },
-    { name: '인턴십 후기', checked: false },
-    { name: 'IT 해외봉사', checked: false },
-  ],
-};
 
 export function MainGridContainer() {
   const searchParams = useSearchParams();
   const query = searchParams.get('query');
 
   const [contents, setContents] = useState<ContentResponseType[]>();
-  const [filters, setFilters] = useState<Filter[]>(
-    groupedFilterList['성대한만남']
-  );
 
-  const filterToggle = (name: CategoryKey) => {
-    setFilters((prev) =>
-      prev.map((filter) =>
-        filter.name === name
-          ? { ...filter, checked: true }
-          : { ...filter, checked: false }
-      )
-    );
-  };
-
-  const filterOn = (name: CategoryKey) => {
-    setFilters((prev) =>
-      prev.map((filter) =>
-        filter.name === name ? { ...filter, checked: true } : filter
-      )
-    );
-  };
+  const { filters, setFilters, filterToggle, filterOn, groupedFilterList } =
+    useFilter();
 
   useEffect(() => {
     const topic = (searchParams.get('topic') ?? '성대한만남') as Topic;
@@ -68,6 +31,7 @@ export function MainGridContainer() {
     setFilters(groupedFilterList[topic]);
 
     filterOn(category);
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [searchParams]);
 
   useEffect(() => {
